pyproject.toml
[project]
name = "db-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=0.8.0",
    "influxdb-client>=1.36.0",
    "matplotlib>=3.5.0",
    "numpy>=1.23.0",
    "openpyxl>=3.0.0",
    "pandas>=1.5.0",
    "plotly>=5.14.0",
    "psycopg2-binary>=2.9.0",
    "seaborn>=0.12.0",
    "streamlit>=1.28.0",
]
